refactoring and putting more functions into utils
