+++ test/components/agents/test_agent.py
@@ -11,12 +11,11 @@ from typing import Any
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
-from jinja2 import TemplateSyntaxError
 from openai import Stream
 from openai.types.chat import ChatCompletionChunk, chat_completion_chunk
 
 from haystack import Document, Pipeline, component, tracing
-from haystack.components.agents.agent import Agent, _accumulate_usage, _select_tools_by_name
+from haystack.components.agents.agent import Agent
 from haystack.components.agents.state import State, merge_lists, replace_values
 from haystack.components.agents.tool_calling import _run_tool
 from haystack.components.builders.chat_prompt_builder import ChatPromptBuilder
@@ -41,10 +40,6 @@ from haystack.utils import Secret, serialize_callable
 
 def _user_msg(text: str) -> str:
     return f'{{% message role="user" %}}{text}{{% endmessage %}}'
-
-
-def _sys_msg(text: str) -> str:
-    return f'{{% message role="system" %}}{text}{{% endmessage %}}'
 
 
 def _assistant_with_usage(text: str | None = None, *, tool_calls=None, usage: dict[str, Any] | None = None):
@@ -1404,51 +1399,6 @@ class TestAgentExitReasonAsync:
         assert result["exit_reason"] == "max_agent_steps"
 
 
-class TestAccumulateUsage:
-    """Unit tests for the `_accumulate_usage` helper used to merge ChatGenerator usage dicts."""
-
-    def test_sums_flat_numeric_keys(self):
-        current = {"prompt_tokens": 10, "completion_tokens": 5, "total_tokens": 15}
-        new = {"prompt_tokens": 3, "completion_tokens": 2, "total_tokens": 5}
-        assert _accumulate_usage(current, new) == {"prompt_tokens": 13, "completion_tokens": 7, "total_tokens": 20}
-
-    def test_merges_nested_detail_dicts_recursively(self):
-        current = {"prompt_tokens": 10, "completion_tokens_details": {"reasoning_tokens": 2, "audio_tokens": 0}}
-        new = {
-            "prompt_tokens": 4,
-            "completion_tokens_details": {"reasoning_tokens": 3, "audio_tokens": 1},
-            "prompt_tokens_details": {"cached_tokens": 6},
-        }
-        assert _accumulate_usage(current, new) == {
-            "prompt_tokens": 14,
-            "completion_tokens_details": {"reasoning_tokens": 5, "audio_tokens": 1},
-            "prompt_tokens_details": {"cached_tokens": 6},
-        }
-
-    def test_adds_keys_missing_in_current(self):
-        assert _accumulate_usage({"prompt_tokens": 5}, {"completion_tokens": 7}) == {
-            "prompt_tokens": 5,
-            "completion_tokens": 7,
-        }
-
-    def test_empty_current_dict_returns_copy_of_new(self):
-        new = {"prompt_tokens": 5, "details": {"cached_tokens": 1}}
-        result = _accumulate_usage({}, new)
-        assert result == new
-        # Nested dicts must be deep-copied so future merges don't mutate the source.
-        new["details"]["cached_tokens"] = 999
-        assert result["details"]["cached_tokens"] == 1
-
-    def test_non_dict_non_numeric_falls_back_to_new(self):
-        # Strings, lists, or any other type that isn't a dict-or-number pair returns `new` unchanged.
-        assert _accumulate_usage("old-model", "new-model") == "new-model"
-        assert _accumulate_usage(5, "stringified") == "stringified"
-        assert _accumulate_usage({"model": "gpt-x"}, {"model": "gpt-y"}) == {"model": "gpt-y"}
-
-    def test_sums_floats(self):
-        assert _accumulate_usage(1.5, 2.25) == 3.75
-
-
 class TestAgentTracing:
     def test_agent_tracing_span_run(self, caplog, monkeypatch, weather_tool):
         chat_generator = MockChatGeneratorWithoutRunAsync()
@@ -1816,55 +1766,6 @@ class TestAgentTracing:
         assert agent_span.parent_span == agent_component_span
 
 
-class TestSelectToolsByName:
-    """Tests for the _select_tools_by_name helper (runtime tool-name selection)."""
-
-    def test_selects_standalone_tools_by_name(self, weather_tool: Tool, component_tool: Tool):
-        result = _select_tools_by_name([weather_tool, component_tool], [weather_tool.name])
-        assert result == [weather_tool]
-
-    def test_raises_on_invalid_name(self, weather_tool: Tool, component_tool: Tool):
-        with pytest.raises(
-            ValueError, match="The following tool names are not valid: {'invalid_tool_name'}. Valid tool names are: ."
-        ):
-            _select_tools_by_name([weather_tool, component_tool], ["invalid_tool_name"])
-
-    def test_raises_when_no_tools_configured(self, weather_tool: Tool):
-        with pytest.raises(ValueError, match="No tools were configured for the Agent at initialization."):
-            _select_tools_by_name([], [weather_tool.name])
-
-    def test_returns_isolated_spawn_with_selection(self, weather_tool: Tool, component_tool: Tool):
-        """A Toolset exposing a requested name is replaced by an isolated spawn carrying the selection.
-
-        The shared, configured Toolset is not mutated.
-        """
-        toolset = Toolset([weather_tool, component_tool])
-
-        result = _select_tools_by_name([toolset], [weather_tool.name])
-
-        assert len(result) == 1
-        spawned = result[0]
-        assert isinstance(spawned, Toolset)
-        assert spawned is not toolset  # an isolated per-run copy
-        assert spawned._selected_tool_names == {weather_tool.name}
-        assert [tool.name for tool in spawned] == [weather_tool.name]
-        # The configured toolset is untouched.
-        assert toolset._selected_tool_names is None
-
-    def test_mixed_standalone_tools_and_toolsets(self, weather_tool: Tool, component_tool: Tool):
-        toolset = Toolset([weather_tool])
-
-        result = _select_tools_by_name([component_tool, toolset], [weather_tool.name, component_tool.name])
-
-        # The standalone tool is passed through; the toolset is replaced by an isolated spawn with the selection.
-        assert component_tool in result
-        spawns = [t for t in result if isinstance(t, Toolset)]
-        assert len(spawns) == 1
-        assert spawns[0] is not toolset
-        assert spawns[0]._selected_tool_names == {weather_tool.name}
-        assert toolset._selected_tool_names is None
-
-
 class TestAgentToolSelection:
     def test_tool_selection_new_tool(self, weather_tool: Tool, component_tool: Tool):
         chat_generator = MockChatGenerator("Hello")
@@ -1993,136 +1894,6 @@ class TestRegisterPromptVariables:
             match="Variable 'streaming_callback' from user_prompt conflicts with input names in the run method.",
         ):
             make_agent(user_prompt=_user_msg("{{streaming_callback}} is already a run parameter."))
-
-
-class TestPrompts:
-    def test_system_prompt_incorrect_jinja2_syntax_raises(self, make_agent):
-        with pytest.raises(TemplateSyntaxError):
-            make_agent(system_prompt="{% message role='system' %}Incomplete syntax.")
-
-    def test_system_prompt_plain_string(self, make_agent):
-        agent = make_agent(system_prompt="You are a helpful assistant.")
-        assert agent._system_chat_prompt_builder is not None
-        result = agent.run(messages=[ChatMessage.from_user("Hi")])
-        assert result["messages"][0].is_from(ChatRole.SYSTEM)
-        assert result["messages"][0].text == "You are a helpful assistant."
-
-    def test_system_prompt_plain_string_with_template_variables(self, make_agent):
-        agent = make_agent(system_prompt="You are an assistant for {{company}}. Your role is {{role}}.")
-        assert agent._system_chat_prompt_builder is not None
-        assert set(agent._system_chat_prompt_builder.variables) == {"company", "role"}
-
-        result = agent.run(messages=[ChatMessage.from_user("Hi")], company="Acme", role="support agent")
-        sys_msg = result["messages"][0]
-        assert sys_msg.is_from(ChatRole.SYSTEM)
-        assert sys_msg.text == "You are an assistant for Acme. Your role is support agent."
-
-        input_names = set(agent.__haystack_input__._sockets_dict.keys())
-        assert "company" in input_names
-        assert "role" in input_names
-
-    def test_system_prompt_with_template_variables(self, make_agent):
-        agent = make_agent(system_prompt=_sys_msg("You are an assistant for {{company}}. Your role is {{role}}."))
-        assert agent._system_chat_prompt_builder is not None
-        assert set(agent._system_chat_prompt_builder.variables) == {"company", "role"}
-
-        result = agent.run(messages=[ChatMessage.from_user("Hi")], company="Acme", role="support agent")
-        sys_msg = result["messages"][0]
-        assert sys_msg.is_from(ChatRole.SYSTEM)
-        assert sys_msg.text == "You are an assistant for Acme. Your role is support agent."
-
-        input_names = set(agent.__haystack_input__._sockets_dict.keys())
-        assert "company" in input_names
-        assert "role" in input_names
-
-    def test_system_prompt_with_meta(self, make_agent):
-        agent = make_agent(
-            system_prompt="{% message role='system' meta={'key': 'value'} %}System message with meta{% endmessage %}"
-        )
-        assert agent._system_chat_prompt_builder is not None
-
-        result = agent.run(messages=[ChatMessage.from_user("Hi")])
-        messages = result["messages"]
-        assert messages[0].is_from(ChatRole.SYSTEM)
-        assert messages[0].text == "System message with meta"
-        assert messages[0].meta == {"key": "value"}
-
-    def test_user_prompt_only_variables_forwarded_to_builder(self, make_agent):
-        agent = make_agent(user_prompt=_user_msg("Question: {{question}}"))
-        # 'irrelevant_kwarg' is not a template variable — must not raise
-        result = agent.run(messages=[], question="Will it snow?", irrelevant_kwarg="unused")
-        assert "messages" in result
-
-    def test_user_prompt_plain_string_with_template_variables(self, make_agent):
-        agent = make_agent(user_prompt="Question: {{question}}")
-        result = agent.run(messages=[], question="Will it snow?")
-        user_messages = [m for m in result["messages"] if m.is_from(ChatRole.USER)]
-        assert user_messages[0].text == "Question: Will it snow?"
-
-        input_names = set(agent.__haystack_input__._sockets_dict.keys())
-        assert "question" in input_names
-
-    def test_user_prompt_with_template_variables(self, make_agent):
-        agent = make_agent(
-            user_prompt=_user_msg(
-                "Hello {{name|upper}}, check weather for: "
-                + "{% for c in cities %}{{c}}{% if not loop.last %}, {% endif %}{% endfor %}"
-                + " on {{date}}?"
-            )
-        )
-        result = agent.run(messages=[], name="Alice", cities=["Berlin", "Paris", "Rome"], date="2024-01-15")
-        user_messages = [m for m in result["messages"] if m.is_from(ChatRole.USER)]
-        assert user_messages[0].text == "Hello ALICE, check weather for: Berlin, Paris, Rome on 2024-01-15?"
-
-        input_names = set(agent.__haystack_input__._sockets_dict.keys())
-        assert "name" in input_names
-        assert "cities" in input_names
-        assert "date" in input_names
-
-    def test_user_prompt_appended_after_initial_messages(self, make_agent):
-        agent = make_agent(user_prompt=_user_msg("And now: {{query}}"))
-        initial_messages = [ChatMessage.from_user("First message")]
-        result = agent.run(messages=initial_messages, query="What is the weather?")
-        user_messages = [m for m in result["messages"] if m.is_from(ChatRole.USER)]
-        assert user_messages[0].text == "First message"
-        assert user_messages[1].text == "And now: What is the weather?"
-
-    def test_system_prompt_and_user_prompt(self, make_agent):
-        agent = make_agent(
-            system_prompt=_sys_msg("You help users of {{project}}."),
-            user_prompt=_user_msg("Tell me about {{topic}} in the {{project}} context."),
-        )
-        assert agent._system_chat_prompt_builder is not None
-        assert agent._user_chat_prompt_builder is not None
-
-        result = agent.run(messages=[], project="Haystack", topic="pipelines")
-        messages = result["messages"]
-        assert messages[0].is_from(ChatRole.SYSTEM)
-        assert messages[0].text == "You help users of Haystack."
-        user_messages = [m for m in messages if m.is_from(ChatRole.USER)]
-        assert user_messages[0].text == "Tell me about pipelines in the Haystack context."
-
-    def test_prompt_wrong_role_raises_at_init(self, make_agent):
-        with pytest.raises(ValueError, match="system_prompt message block must have role 'system'"):
-            make_agent(system_prompt=_user_msg("This is a user message, not system."))
-
-        with pytest.raises(ValueError, match="user_prompt message block must have role 'user'"):
-            make_agent(user_prompt=_sys_msg("This is a system message, not user."))
-
-    def test_dynamic_prompt_role_raises_at_runtime(self, make_agent):
-        agent = make_agent(user_prompt="{% message role=role_name %}Question: {{question}}{% endmessage %}")
-        with pytest.raises(ValueError, match="user_prompt must render to a user message"):
-            agent.run(messages=[], role_name="assistant", question="Will it snow?")
-
-    def test_prompt_multiple_message_blocks_raises_at_init(self, make_agent):
-        multi_message_prompt = """{% message role='system' %}You are a helpful assistant.{% endmessage %}
-        {% message role='user' %}How are you?{% endmessage %}"""
-
-        with pytest.raises(ValueError, match="system_prompt must define exactly one message block"):
-            make_agent(system_prompt=multi_message_prompt)
-
-        with pytest.raises(ValueError, match="user_prompt must define exactly one message block"):
-            make_agent(user_prompt=multi_message_prompt)
 
 
 @pytest.mark.integration
